requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
aiogram==3.4.1
python-telegram-bot==20.8
sentry-sdk==1.40.4
pytest==8.0.2
pytest-django==4.8.0
pytest-asyncio==0.23.5
black==24.2.0
isort==5.13.2
flake8==7.0.0
Pillow==10.2.0 